[refactor♻️] 인기도 레디스 조회/저장 서비스 -> 퍼시스턴스 계층으로 이동
